Create Linear Regression Analysis Requirements
INTRODUCTION
Managers are required to organize, interpret, and display data that is relevant to the real-world business decisions they must make in their businesses, and these business decisions must be based on relevant and reliable data. The use of analytical tools will improve your ability to use data to make these informed decisions.



In this task, you will address the business situation in the attached "Linear Regression Analysis Resources" scenario. You will access the scenario and data set by entering your student ID number in the “Start” tab of the attachment, then continuing to the “Scenario” tab. Using this data set, you will perform a linear regression analysis and write a report in which you recommend a solution by summarizing the key details of your analysis.



For full functionality of the scenario and data set attachment, you are strongly encouraged to use Microsoft Excel, which is available via the Microsoft Office 365 subscription service provided to all WGU students. It can be downloaded using the "Microsoft Office 365" link in the weblinks section.

SCENARIO
Refer to the scenario located in the attached “Linear Regression Analysis Resources.”

REQUIREMENTS
Your submission must be your original work. No more than a combined total of 30% of the submission and no more than a 10% match to any one individual source can be directly quoted or closely paraphrased from sources, even if cited correctly. The originality report that is provided when you submit your task can be used as a guide.

 

You must use the rubric to direct the creation of your submission because it provides detailed criteria that will be used to evaluate your work. Each requirement below may be evaluated by more than one rubric aspect. The rubric aspect titles may contain hyperlinks to relevant portions of the course.

 

Tasks may not be submitted as cloud links, such as links to Google Docs, Google Slides, OneDrive, etc., unless specified in the task requirements. All other submissions must be file types that are uploaded and submitted as attachments (e.g., .docx, .pdf, .ppt, .xls).

 

Complete your linear regression analysis and create a report (suggested length of 2–4 pages or 800 words) by doing the following:

 

Note: You are encouraged to use the template located within the attached “Linear Regression Analysis Resources” to complete your analysis. While it is required that you use the scenario and data set located in the attachment, the use of the data analysis template is optional.

 

A.  Describe a business question that could be answered by applying linear regression analysis and is derived from the scenario in the attached “Linear Regression Analysis Resources."

 

B.  Describe the data provided in the attached “Linear Regression Analysis Resources” by doing the following:

1.  Describe the relevant data characteristics for your linear regression analysis, including each of the following:

•   the independent variable(s)

•   the dependent variable

•   type of data 

•   quantity of data

2.  Create a graphical display of the data using a scatter plot or line chart, including each of the following:

•   chart title

•   legend

•   axis titles

•   data intervals

 

Note: This display should be a summary or representation of the data provided, not raw data.

 

C.  Report how you analyzed the data using linear regression by doing the following:

1.  Provide the output and calculations of the linear regression analysis you performed.

 

Note: You may submit the analysis output and calculations using a separate spreadsheet attachment or the optional template in the attached “Linear Regression Analysis Resources.”

 

Note: The output should include the output from the software you used to perform the analysis. Refer to “Prepare for the Performance Assessment Task 1” in the course of study to see examples of acceptable output.

 

2.  Justify why linear regression is the appropriate analysis technique for predicting the dependent variable, including relevant details from the scenario to support your justification.

 

D.  Describe the implications of your data analysis from the scenario by doing the following:

1.  State the null hypothesis for this linear regression analysis.

2.  Interpret the results of the data analysis by doing the following:

a.  Discuss the goodness of fit with the supporting test statistic from your linear regression analysis output.

b.  Discuss the significance of the independent variable(s) with support from your linear regression analysis results.

c.  Create the linear equation and explain its purpose using your analysis results. 

3.  Discuss a limitation of the research that could affect a recommended course of action.

4.  Recommend a course of action that aligns with your linear regression analysis results.

 

Note: Your recommendation should focus on the results of your analytic technique output from part C1.

 

E.  Acknowledge sources, using in-text citations and references, for content that is quoted, paraphrased, or summarized.

 

F.  Demonstrate professional communication in the content and presentation of your submission.