Create Decision Tree Analysis Requirements
INTRODUCTION
Managers are required to organize, interpret, and display data that is relevant to the real-world business decisions they must make in their businesses. Business decisions must be based on relevant and reliable data. The use of analytical tools will improve your ability to use data to make informed decisions.

In this task, you will address the business situation in the attached scenario. You will access the scenario and data set by entering your student ID number in the “Start” tab of the “Decision Tree Resources” attachment. The scenario and data set are located in the “Decision Tree Scenario” tab. Using this data set, you will perform a decision tree analysis and recommend a solution. This recommendation will be included in a report that you will write, summarizing the key details of your analysis.

For full functionality of the scenario and data attachment, you are strongly encouraged to use Microsoft Excel, which is available via the Microsoft Office 365 subscription service provided to all WGU students. It can be downloaded using the "Microsoft Office 365" link in the weblinks section.
SCENARIO
Refer to the scenario located in the attached “Decision Tree Analysis Resources.”
REQUIREMENTS
Your submission must be your original work. No more than a combined total of 30% of the submission and no more than a 10% match to any one individual source can be directly quoted or closely paraphrased from sources, even if cited correctly. The originality report that is provided when you submit your task can be used as a guide.

 

You must use the rubric to direct the creation of your submission because it provides detailed criteria that will be used to evaluate your work. Each requirement below may be evaluated by more than one rubric aspect. The rubric aspect titles may contain hyperlinks to relevant portions of the course.

 

Tasks may not be submitted as cloud links, such as links to Google Docs, Google Slides, OneDrive, etc., unless specified in the task requirements. All other submissions must be file types that are uploaded and submitted as attachments (e.g., .docx, .pdf, .ppt).

 

Complete your decision tree analysis and create a report (suggested length of 2–4 pages or 800 words) by doing the following:

 

Note: You are encouraged to use the template located within the attached “Decision Tree Analysis Resources” to complete your analysis. While it is required that you use the scenario and data set located in the attachment, the use of the data analysis template is optional.

 

A.  Describe a business question that could be answered by applying decision tree analysis and is derived from the scenario in the attached “Decision Tree Analysis Resources."

 

B.  Identify the relevant data values required for your decision tree analysis, including the following:

•   probabilities

•   payoffs

•   profits

•   demand

 

C.  Report how you analyzed the data using decision tree analysis by doing the following:

1.  Complete a decision tree diagram, including each of the following: 

•   state-of-nature nodes

•   calculated payoffs, each expressed out to two decimal places

•   expected values, each expressed out to two decimal places

 

Note: You can submit the completed decision tree diagram using a separate attachment or the optional template on the attached “Decision Tree Analysis Resources.”

 

Note: Refer to “Prepare for the Performance Assessment Task 2” in the course of study to see examples of acceptable output. 

 

2.  Justify why decision tree analysis is the appropriate analysis technique, including relevant details from the scenario to support your justification.

 

D.  Summarize the implications of your decision tree analysis by doing the following:

1.  Explain the role of probabilities and the role of demand for each branch.

2.  Explain how the expected value of each node is determined based on payoffs.

 

Note: Refer “Prepare for the Performance Assessment Task 2” in the course of study to see an example of an acceptable discussion of results.

 

3.  Discuss one limitation of each of the following:

•   the data elements

•   the decision tree analysis

 

E.  Recommend a course of action that addresses the business question from part A and is based on the results of your decision tree analysis.

 

F.  Acknowledge sources, using in-text citations and references, for content that is quoted, paraphrased, or summarized.

 

G.  Demonstrate professional communication in the content and presentation of your submission.